pyproject.toml

[project]
name = "home-maintenance-tracker"
version = "0.1.0"
description = "Local-first Home Maintenance Tracker (Flask + SQLite)"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Tk" }]

dependencies = [
  "flask>=3.0.0",
  "flask-login>=0.6.3",
  "flask_sqlalchemy>=3.1.1",
  "flask-migrate>=4.0.7",
  "flask-wtf>=1.2.1",
  "email-validator>=2.2.0",
  "python-dotenv>=1.0.1",
  "gunicorn>=21.2.0",
  "alembic>=1.13.1",
  "wtforms>=3.1.2",
  "pillow>=10.4.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
